feat!: * add dpop as proof in access token * change access token 'typ' to 'at+jwt' * add issuer claim in access token matching dpop 'sub' claim and duplicate it in 'client_id' * add hardcoded 'scope' & 'api_version' * fixed ClientId format & client part changed from u16 to u64 * dpop token expiry is configurable

+++ jwt-ffi/src/haskell.rs
@@ -15,7 +15,7 @@ impl RustyJwtToolsFfi {
     pub extern "C" fn generate_dpop_access_token(
         dpop_proof: *const c_char,
         user: *const c_char,
-        client_id: u16,
+        client_id: u64,
         domain: *const c_char,
         backend_nonce: *const c_char,
         uri: *const c_char,
